🚚 Improve organization and naming consistency
