+++ src/composables/useCvss4Calculator.ts
